reorg tests to make it easier to read
